refactor: directory composition.
