+++ main.go
@@ -4,7 +4,7 @@ Copyright © 2023 at0x0ft <26642966+at0x0ft@users.noreply.github.com>
 */
 package main
 
-import "github.com/at0x0ft/pathru/internal/app/pathru/cmd"
+import "github.com/at0x0ft/pathru/pkg/cmd"
 
 func main() {
     cmd.Execute()
